requirements.txt
Flask
googletrans==4.0.0-rc1
SpeechRecognition
gtts
pygame
pydub
python-dotenv
